feat: enhance logic of checking and saving of the options argument
Add more checks to the `checkOptions` to verify the options argument's validity.
Create a seperate `setOptions` method to save the options tag for the future use,
as well as it is appointed to set default values for the unnecessary fields
(description, fields' isRequired, validationFunctions, errorMessage, attributes).
Refactor the subsequent code in order to remove needless checks.